feat: fix route structure
